+++ onadata/apps/api/tasks.py
@@ -6,7 +6,6 @@ import os
 import sys
 import logging
 from datetime import timedelta
-from typing import Optional
 
 from celery.result import AsyncResult
 from django.conf import settings
@@ -21,7 +20,6 @@ from onadata.libs.utils.email import send_generic_email
 from onadata.libs.utils.model_tools import queryset_iterator
 from onadata.apps.logger.models import Instance, ProjectInvitation, XForm
 from onadata.libs.utils.email import ProjectInvitationEmail
-from onadata.libs.utils.user_auth import accept_project_invitation
 from onadata.celeryapp import app
 
 User = get_user_model()
@@ -147,24 +145,3 @@ def send_project_invitation_email_async(
     else:
         email = ProjectInvitationEmail(invitation, url)
         email.send()
-
-
-@app.task()
-def accept_project_invitation_async(user_id: str, invitation_id: Optional[str] = None):
-    """Accpet project invitation asynchronously"""
-    invitation = None
-
-    if invitation_id:
-        try:
-            invitation = ProjectInvitation.objects.get(id=invitation_id)
-        except ProjectInvitation.DoesNotExist as err:
-            logging.exception(err)
-
-    try:
-        user = User.objects.get(pk=user_id)
-
-    except User.DoesNotExist as err:
-        logging.exception(err)
-
-    else:
-        accept_project_invitation(user, invitation)
