+++ onadata/apps/api/viewsets/user_profile_viewset.py
@@ -206,18 +206,6 @@ class UserProfileViewSet(
 
         return obj
 
-    def get_serializer(self, *args, **kwargs):
-        """Override get_serializer"""
-        if (invitation_id := self.request.query_params.get("invitation_id")) and (
-            invitation_token := self.request.query_params.get("invitation_token")
-        ):
-            draft_request_data = self.request.data.copy()
-            draft_request_data["invitation_id"] = invitation_id
-            draft_request_data["invitation_token"] = invitation_token
-            kwargs["data"] = draft_request_data
-
-        return super().get_serializer(*args, **kwargs)
-
     def update(self, request, *args, **kwargs):
         """Update user in cache and db"""
         username = kwargs.get("user")
